+++ agarwalamit/src/main/java/playground/agarwalamit/fundamentalDiagrams/core/BinaryAdditionModule.java
@@ -1,4 +1,4 @@
-package playground.agarwalamit.fundamentalDiagrams;
+package playground.agarwalamit.fundamentalDiagrams.core;
 
 import java.util.List;
