+++ agarwalamit/src/main/java/playground/agarwalamit/fundamentalDiagrams/core/FDNetworkGenerator.java
@@ -17,7 +17,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package playground.agarwalamit.fundamentalDiagrams;
+package playground.agarwalamit.fundamentalDiagrams.core;
 
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
