+++ agarwalamit/src/main/java/playground/agarwalamit/fundamentalDiagrams/core/TravelModesFlowDynamicsUpdator.java
@@ -18,7 +18,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package playground.agarwalamit.fundamentalDiagrams;
+package playground.agarwalamit.fundamentalDiagrams.core;
 
 import java.util.LinkedList;
 import java.util.List;
