+++ agarwalamit/src/main/java/playground/agarwalamit/fundamentalDiagrams/dynamicPCU/RunRaceTrackDynamicPCUExample.java
@@ -27,7 +27,8 @@ import org.matsim.core.scenario.ScenarioUtils;
 import org.matsim.vehicles.VehicleType;
 import org.matsim.vehicles.VehicleUtils;
 import org.matsim.vehicles.Vehicles;
-import playground.agarwalamit.fundamentalDiagrams.FundamentalDiagramDataGenerator;
+
+import playground.agarwalamit.fundamentalDiagrams.core.FundamentalDiagramDataGenerator;
 import playground.agarwalamit.fundamentalDiagrams.dynamicPCU.projectedArea.VehicleProjectedAreaMarker;
 import playground.agarwalamit.fundamentalDiagrams.dynamicPCU.projectedArea.VehicleProjectedAreaRatio;
 import playground.agarwalamit.utils.FileUtils;
